requirements.txt
anyio==4.9.0
bcrypt==4.3.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
cssselect==1.3.0
cssutils==2.11.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
h11==0.16.0
idna==3.10
lxml==5.4.0
more-itertools==10.7.0
passlib==1.7.4
pillow==11.2.1
premailer==3.10.0
pyasn1==0.4.8
pycparser==2.22
pydantic==1.10.22
pymongo==3.12.0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
qrcode==8.1
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
yagmail==0.15.293

